added some entities, controllers etc

+++ web_app/src/main/java/com/example/web_app/repositories/PrescriptionsRepository.java
@@ -1,12 +1,13 @@
 package com.example.web_app.repositories;
 
+import com.example.web_app.entity.Prescription;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface PrescriptionsRepository extends JpaRepository<Prescription, Integer> {
-    List<Prescription> findPrescriptionsByDoctorId(int id);
-    List<Prescription> findPrescriptionsByPatientId(int id);
+public interface PrescriptionsRepository extends JpaRepository<Prescription, Long> {
+    List<Prescription> findPrescriptionsByDoctorId(Long id);
+    List<Prescription> findPrescriptionsByPatientId(Long id);
 }
